+++ assets/src/components/resource/editors/AddResourceOrDropTarget.tsx
@@ -1,15 +1,15 @@
-import React from 'react';
-import { ActivityEditorMap } from 'data/content/editors';
-import { ResourceContext, ResourceContent } from 'data/content/resource';
-import { Objective } from 'data/content/objective';
-import { ResourceId } from 'data/types';
-import { AddResourceContent } from 'components/content/add_resource_content/AddResourceContent';
-import * as Immutable from 'immutable';
-import { DropTarget } from './dragndrop/DropTarget';
-import { ActivityEditContext } from 'data/content/activity';
-import { AddContent } from 'components/content/add_resource_content/AddContent';
 import { AddActivity } from 'components/content/add_resource_content/AddActivity';
+import { AddContent } from 'components/content/add_resource_content/AddContent';
 import { AddOther } from 'components/content/add_resource_content/AddOther';
+import { AddResourceContent } from 'components/content/add_resource_content/AddResourceContent';
+import { ActivityEditContext } from 'data/content/activity';
+import { ActivityEditorMap } from 'data/content/editors';
+import { Objective } from 'data/content/objective';
+import { ResourceContent, ResourceContext } from 'data/content/resource';
+import { ResourceId } from 'data/types';
+import * as Immutable from 'immutable';
+import React from 'react';
+import { DropTarget } from './dragndrop/DropTarget';
 
 export type AddResourceOrDropTargetProps = {
   isReorderMode: boolean;
@@ -33,9 +33,7 @@ export const AddResourceOrDropTarget = (props: AddResourceOrDropTargetProps) => 
   return (
     <AddResourceContent {...props} isLast={props.id === 'last'}>
       <AddContent {...props} />
-      <hr />
       <AddActivity {...props} />
-      <hr />
       <AddOther {...props} />
     </AddResourceContent>
   );