+++ lib/oli_web/live/products/payments_view.ex
@@ -80,7 +80,7 @@ defmodule OliWeb.Products.PaymentsView do
     ~F"""
     <div>
 
-      <CreateCodes count={@code_count} product_slug={@product_slug} click="create_codes" change="change_count"/>
+      <CreateCodes id="create_codes" count={@code_count} product_slug={@product_slug} click="create_codes" change="change_count"/>
 
       <hr class="mt-5 mb-5"/>