+++ lib/oli_web/live/curriculum/entries/details.ex
@@ -4,16 +4,15 @@ defmodule OliWeb.Curriculum.DetailsLive do
   """
 
   use OliWeb, :live_component
-  import OliWeb.Projects.Table, only: [time_ago: 2]
 
   def render(assigns) do
     ~L"""
     <div class="entry-section d-flex flex-column col-4">
       <small class="text-muted">
-        Created <%= time_ago(assigns, @child.resource.inserted_at) %>
+        Created <%= Timex.format!(@child.resource.inserted_at, "{relative}", :relative) %>
       </small>
       <small class="text-muted">
-        Updated <%= time_ago(assigns, @child.inserted_at) %> by <%= @child.author.name %>
+        Updated <%= Timex.format!(@child.inserted_at, "{relative}", :relative) %> by <%= @child.author.name %>
       </small>
     </div>
     """