+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 ### Bug Fixes
 
 - Fix a style issue with the workspace footer
+- Prevent objectives used in selections from being deleted
 - Fix an issue where modals misbehaved sporadically
+- Move "Many Students Wonder" from activity styling to content styling
+- Fix an issue where nonstructural section resources were missing after update
 
 ### Enhancements
 
@@ -14,6 +17,10 @@
 - Add advanced section creation remix
 - Allow for section creation from course products
 - Add analytics / insights data export button
+- Add ability for an admin to browse all course sections
+- Add server driven paged, sortable table for project list
+- Add ability to remix materials from multiple projects
+- Fix insert content popup in page editor
 
 ### Release Notes